migrate from sqlite to postgres v2, update application.yaml and sql script

+++ src/main/resources/db/migration/V001__create_table_users.sql
@@ -1,6 +1,6 @@
 CREATE TABLE users
 (
-    id             INTEGER PRIMARY KEY,
+    id             BIGSERIAL PRIMARY KEY,
     username       VARCHAR(300) UNIQUE NOT NULL,
     name           VARCHAR(300) NOT NULL,
     register_date  VARCHAR(100) NOT NULL,
